requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask
streamlit
ensure==1.0.2
python-box==6.0.2
PyYAML
datasets

-e .